V4 finished

Closes #4

See merge request isc/2022-23/niveau-1/1242.2-langagecpp/isc1ab/goloviatinski_fil_rouge!6

+++ include/hero/class/Rogue.h
@@ -5,7 +5,7 @@
 namespace He_ARC::rpg {
     class Rogue : public Hero {
         public:
-            Rogue(int=5, int=10, int=7, double=100, Weapon *weapon=new Weapon(100,"Rusty Knife"), string="Locke");
+            Rogue(int=5, int=10, int=7, int=100, Weapon *weapon=new Weapon(100, "Rusty Knife"), string="Locke");
             void interact(const Hero&) override;
     };
 }
