+++ src/hero/class/Rogue.cpp
@@ -1,7 +1,7 @@
 #include "class/Rogue.h"
 
 namespace He_ARC::rpg {
-    Rogue::Rogue(int _strength, int _agility, int _intelligence, double _hp, Weapon *_weapon, string _name) : 
+    Rogue::Rogue(int _strength, int _agility, int _intelligence, int _hp, Weapon *_weapon, string _name) : 
     Hero(_strength, _agility, _intelligence, _hp, _weapon, _name) {}
