+++ src/hero/class/Warrior.cpp
@@ -1,7 +1,7 @@
 #include "class/Warrior.h"
 
 namespace He_ARC::rpg {
-    Warrior::Warrior(int _strength, int _agility, int _intelligence, double _hp, Weapon *_weapon, string _name) : 
+    Warrior::Warrior(int _strength, int _agility, int _intelligence, int _hp, Weapon *_weapon, string _name) : 
     Hero(_strength, _agility, _intelligence, _hp, _weapon, _name) {}
