+++ src/hero/class/wizard/Necromancer.cpp
@@ -1,7 +1,7 @@
 #include "class/wizard/Necromancer.h"
 
 namespace He_ARC::rpg {
-    Necromancer::Necromancer(int _strength, int _agility, int _intelligence, double _hp, int _mana, Weapon *_weapon, string _name) : 
+    Necromancer::Necromancer(int _strength, int _agility, int _intelligence, int _hp, int _mana, Weapon *_weapon, string _name) : 
     Wizard(_strength, _agility, _intelligence, _hp, _mana, _weapon, _name) {}
